python/pgbadger-cronjob/docker/requirements.txt
oci
pyyaml
